Consolidated typography and labelled text
