+++ src/app/ui/android/Typography.tsx
@@ -1,4 +1,17 @@
-import styles from "./Text.module.css";
+import styles from "./Typography.module.css";
+
+interface LabelProps {
+    text: string;
+    color?: 'primary' | 'white';
+}
+
+export function Label({ text, color = 'primary' }: LabelProps) {
+    const colorClass = color === 'white' ? styles['label-white']: styles['label-primary'];
+    
+    return(
+        <p className={`${styles.label} ${colorClass}`}>{text}</p>
+    );
+}
 
 interface TextProps {
     text: string;
@@ -7,7 +20,7 @@ interface TextProps {
     color?: 'black' | 'success' | 'error' | 'white';
 }
 
-export default function Text({ text, fill=false, size = 'regular', color = 'black' }: TextProps) {
+export function Text({ text, fill=false, size = 'regular', color = 'black' }: TextProps) {
     const fillClass = fill ? styles['text-fill'] : '';
     const sizeClass = size === 'large' ? styles['text-large']: styles['text-regular'];
     const colorClass = color === 'success'
@@ -22,5 +35,11 @@ export default function Text({ text, fill=false, size = 'regular', color = 'blac
         <p className={`${fillClass} ${colorClass} ${sizeClass}`}>
             {text}
         </p>
+    );
+}
+
+export function PageTitle({ text } : { text: string }) {
+    return(
+        <p className={styles.pagetitle}>{text}</p>
     );
 }
