--- src/app/ui/android/typography/Label.module.css
@@ -1,12 +1,0 @@
-.label {
-    font-size: 0.75rem;
-    font-weight: 700;
-}
-
-.label-primary {
-    color: var(--MC-Primary);
-}
-
-.label-white {
-    color: var(--MC-White);
-}
