--- src/app/ui/android/typography/Label.tsx
@@ -1,14 +1,0 @@
-import styles from "./Label.module.css";
-
-interface LabelProps {
-    text: string;
-    color?: 'primary' | 'white';
-}
-
-export default function Label({ text, color = 'primary' }: LabelProps) {
-    const colorClass = color === 'white' ? styles['label-white']: styles['label-primary'];
-    
-    return(
-        <p className={`${styles.label} ${colorClass}`}>{text}</p>
-    );
-}
