--- src/app/ui/android/typography/LabelledText.module.css
@@ -1,5 +1,0 @@
-.labelledtext {
-    display: flex;
-    flex-direction: column;
-    width: 100%
-}
