--- src/app/ui/android/typography/LabelledText.tsx
@@ -1,14 +1,0 @@
-import styles from "./LabelledText.module.css"
-
-interface LabelledTextProps {
-    gridStyles: React.CSSProperties;
-    children: React.ReactNode;
-}
-
-export default function LabelledText({ gridStyles, children }: LabelledTextProps) {
-    return(
-        <div style={gridStyles} className={styles.labelledtext}>
-            {children}
-        </div>
-    );
-}
