--- src/app/ui/android/typography/PageTitle.module.css
@@ -1,5 +1,0 @@
-.pagetitle {
-    font-size: 1.25rem;
-    font-weight: 500;
-    color: var(--MC-White);
-}
