--- src/app/ui/android/typography/PageTitle.tsx
@@ -1,7 +1,0 @@
-import styles from "./PageTitle.module.css";
-
-export default function PageTitle({ text } : { text: string }) {
-    return(
-        <p className={styles.pagetitle}>{text}</p>
-    );
-}
