+++ src/app/ui/loadflight/LastScan.tsx
@@ -1,8 +1,6 @@
 import styles from "./LastScan.module.css";
-import LabelledTextWithIcon from "../android/LabelledTextWithIcon";
-import LabelledText from "../android/typography/LabelledText";
-import Label from "../android/typography/Label";
-import Text from "../android/typography/Text";
+import LabelledText from "../android/LabelledText";
+import { Label, Text } from "../android/Typography";
 import { CircleCheckIcon, CircleAlertIcon } from "../android/Icons";
 import InstructionalText from "../android/InstructionalText";
 
@@ -20,95 +18,79 @@ export default function LastScan({state='initial'}: LastScanProps) {
         />;
     const successContent = 
         <div className={styles.labelledtextrows}>
-            <LabelledTextWithIcon 
+            <LabelledText 
                 gridStyles={{}}
                 icon={circleCheckIcon}
                 largeText={true}
             >
-                <LabelledText gridStyles={{}}>
-                    <Label text="TAG NUMBER" />
-                    <Text text="0105123456" color="success" size="large" fill={true} />
-                </LabelledText>
-            </LabelledTextWithIcon>
-            <LabelledTextWithIcon 
+                <Label text="TAG NUMBER" />
+                <Text text="0105123456" color="success" size="large" fill={true} />
+            </LabelledText>
+            <LabelledText 
                 gridStyles={{}}
                 icon={null}
             >
-                <LabelledText gridStyles={{}}>
-                    <Label text="PASSENGER NAME" />
-                    <Text text="CLEMENT/ERNIE" color="success" size="large" fill={true} />
-                </LabelledText>
-            </LabelledTextWithIcon>
+                <Label text="PASSENGER NAME" />
+                <Text text="CLEMENT/ERNIE" color="success" size="large" fill={true} />
+            </LabelledText>
             <div className={styles.baginfo}>
-                <LabelledTextWithIcon 
+                <LabelledText 
                     gridStyles={{width: '55%'}}
                     icon={null}
                 >
-                    <LabelledText gridStyles={{}}>
-                        <Label text="TYPE" />
-                        <Text text="CHECKED BAG" color="success" fill={true} />
-                    </LabelledText>
-                </LabelledTextWithIcon>
-                <LabelledText gridStyles={{width: '45%'}}>
+                    <Label text="TYPE" />
+                    <Text text="CHECKED BAG" color="success" fill={true} />
+                </LabelledText>
+                <LabelledText gridStyles={{width: '45%'}} showIcon={false} icon={null}>
                     <Label text="WEIGHT" />
                     <Text text="STANDARD" color="success" fill={true} />
                 </LabelledText>
             </div>
-            <LabelledTextWithIcon 
+            <LabelledText 
                 gridStyles={{}}
                 icon={null}
             >
-                <LabelledText gridStyles={{}}>
-                    <Label text="RESULT" />
-                    <Text text="Successfully loaded" color="success" fill={true} />
-                </LabelledText>
-            </LabelledTextWithIcon> 
+                <Label text="RESULT" />
+                <Text text="Successfully loaded" color="success" fill={true} />
+            </LabelledText> 
         </div>
     const exceptionContent = 
         <div className={styles.labelledtextrows}>
-            <LabelledTextWithIcon 
+            <LabelledText 
                 gridStyles={{}}
                 icon={circleAlertIcon}
                 largeText={true}
             >
-                <LabelledText gridStyles={{}}>
-                    <Label text="TAG NUMBER" />
-                    <Text text="0105123456" color="error" size="large" fill={true} />
-                </LabelledText>
-            </LabelledTextWithIcon>
-            <LabelledTextWithIcon 
+                <Label text="TAG NUMBER" />
+                <Text text="0105123456" color="error" size="large" fill={true} />
+            </LabelledText>
+            <LabelledText 
                 gridStyles={{}}
                 icon={null}
             >
-                <LabelledText gridStyles={{}}>
-                    <Label text="PASSENGER NAME" />
-                    <Text text="CLEMENT/ERNIE" color="error" size="large" fill={true} />
-                </LabelledText>
-            </LabelledTextWithIcon>
+                <Label text="PASSENGER NAME" />
+                <Text text="CLEMENT/ERNIE" color="error" size="large" fill={true} />
+            </LabelledText>
             <div className={styles.baginfo}>
-                <LabelledTextWithIcon 
+                <LabelledText 
                     gridStyles={{width: '55%'}}
                     icon={null}
                 >
-                    <LabelledText gridStyles={{}}>
-                        <Label text="TYPE" />
-                        <Text text="CHECKED BAG" color="error" fill={true} />
-                    </LabelledText>
-                </LabelledTextWithIcon>
-                <LabelledText gridStyles={{width: '45%'}}>
+                    <Label text="TYPE" />
+                    <Text text="CHECKED BAG" color="error" fill={true} />
+                </LabelledText>
+                <LabelledText gridStyles={{width: '45%'}} showIcon={false} icon={null}>
                     <Label text="WEIGHT" />
                     <Text text="STANDARD" color="error" fill={true} />
                 </LabelledText>
             </div>
-            <LabelledTextWithIcon 
+            <LabelledText 
                 gridStyles={{}}
                 icon={null}
             >
-                <LabelledText gridStyles={{}}>
-                    <Label text="RESULT" />
-                    <Text text="Bag not loaded" color="error" fill={true} />
-                </LabelledText>
-            </LabelledTextWithIcon> 
+                <Label text="RESULT" />
+                <Text text="Bag not loaded" color="error" fill={true} />
+            </LabelledText> 
         </div>
 
     const lastScanContent = state === 'success'
